requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
langgraph>=0.1.0
duckduckgo-search>=6.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
